move nodemailer configs to form router
